+++ src/test/java/eureca/capstone/project/admin/report/service/ReportServiceTest.java
@@ -216,63 +216,6 @@ class ReportServiceTest {
         verify(reportHistoryRepository).findByCriteria(statusCode, keyword, pageable);
     }
 
-    @Test
-    @DisplayName("제재 내역 목록 조회_조건 없음_성공")
-    void getRestrictionList_all_Success() {
-        // given
-        String statusCode = null;
-        String keyword = null;
-        Page<RestrictionTarget> page = new PageImpl<>(List.of(restriction1, restriction2));
-
-        when(restrictionTargetRepository.findByCriteria(statusCode, keyword, pageable)).thenReturn(page);
-
-        // when
-        Page<RestrictionDto> result = reportService.getRestrictionListByStatusCode(statusCode,keyword, pageable);
-
-        // then
-        assertNotNull(result);
-        assertEquals(2, result.getTotalElements());
-        verify(restrictionTargetRepository).findByCriteria(statusCode, keyword, pageable);
-    }
-
-    @Test
-    @DisplayName("제재 내역 목록 조회_상태코드로 필터링_성공")
-    void getRestrictionList_filtered_Success() {
-        // given
-        String statusCode = "PENDING";
-        String keyword = null;
-        Page<RestrictionTarget> page = new PageImpl<>(List.of(restriction1));
-
-        when(restrictionTargetRepository.findByCriteria(statusCode,keyword, pageable)).thenReturn(page);
-
-        // when
-        Page<RestrictionDto> result = reportService.getRestrictionListByStatusCode(statusCode, keyword, pageable);
-
-        // then
-        assertNotNull(result);
-        assertEquals(1, result.getTotalElements());
-        assertEquals("제재 대기중", result.getContent().get(0).getStatus());
-        verify(restrictionTargetRepository).findByCriteria(statusCode,keyword, pageable);
-    }
-
-    @Test
-    @DisplayName("제재 내역 목록 조회_키워드로 검색_성공")
-    void getRestrictionList_SearchByKeyword_Success() {
-        // given
-        String statusCode = null;
-        String keyword = "reporter";
-        Page<RestrictionTarget> page = new PageImpl<>(List.of(restriction1));
-
-        when(restrictionTargetRepository.findByCriteria(statusCode,keyword, pageable)).thenReturn(page);
-
-        // when
-        Page<RestrictionDto> result = reportService.getRestrictionListByStatusCode(statusCode, keyword, pageable);
-
-        assertNotNull(result);
-        assertEquals(1, result.getTotalElements());
-        assertEquals(user1.getEmail(), result.getContent().get(0).getUserEmail());
-        verify(restrictionTargetRepository).findByCriteria(statusCode, keyword, pageable);
-    }
 
 
 
@@ -386,210 +329,16 @@ class ReportServiceTest {
                 .build();
         List<Status> acceptedStatuses = List.of(aiAccept, adminAccept);
 
+        when(restrictionTypeRepository.findById(1L)).thenReturn(Optional.of(restrictionType));
         when(statusManager.getStatus("RESTRICTION", "PENDING"))
                 .thenReturn(pendingStatus);
 
         // when
-        ReflectionTestUtils.invokeMethod(reportService,"applyRestriction", user1, reportType, restrictionType, acceptedStatuses);
+        ReflectionTestUtils.invokeMethod(reportService,"applyRestriction", user1, reportType, 1L, acceptedStatuses);
 
         // then
         verify(restrictionTargetRepository).save(any(RestrictionTarget.class));
-    }
-
-    @DisplayName("제재 승인_영구정지일 경우 사용자 상태만 변경")
-    @Test
-    void acceptRestrictions_Restriction_Success() {
-        // given
-        Status pendingStatus = Status.builder().domain("RESTRICTION").code("PENDING").build();
-        RestrictionTarget restrictionTarget = RestrictionTarget.builder()
-                .restrictionType(RestrictionType.builder().duration(-1).build())
-                .status(pendingStatus)
-                .user(user1)
-                .build();
-
-        ReflectionTestUtils.setField(restrictionTarget, "restrictionTargetId", 1L);
-
-        Status bannedStatus = Status.builder().domain("USER").code("BANNED").build();
-        Status completedStatus = Status.builder().domain("RESTRICTION").code("COMPLETED").build();
-        Status reportCompletedStatus = Status.builder().domain("REPORT").code("COMPLETED").build();
-        Status blurredStatus = Status.builder().domain("FEED").code("BLURRED").build();
-
-        when(statusManager.getStatus("USER", "BANNED")).thenReturn(bannedStatus);
-        when(statusManager.getStatus("RESTRICTION", "COMPLETED")).thenReturn(completedStatus);
-        when(statusManager.getStatus("REPORT", "COMPLETED")).thenReturn(reportCompletedStatus);
-        when(statusManager.getStatus("FEED", "BLURRED")).thenReturn(blurredStatus);
-        when(restrictionTargetRepository.findById(1L)).thenReturn(Optional.of(restrictionTarget));
-        when(reportHistoryRepository.findByRestrictionTarget(restrictionTarget)).thenReturn(List.of(report1));
-
-        // when
-        reportService.acceptRestrictions(1L);
-
-        // then
-        assertEquals(bannedStatus, restrictionTarget.getUser().getStatus());
-        assertNull(restrictionTarget.getExpiresAt());
-        assertEquals(completedStatus, restrictionTarget.getStatus());
-        assertEquals(reportCompletedStatus, report1.getStatus());
-        assertEquals(blurredStatus, report1.getTransactionFeed().getStatus());
-        verify(transactionFeedRepository).saveAll(anyList());
-    }
-
-    @DisplayName("제재 승인_userAuthority에 제재 권한 내역 없는 경우 신규 등록")
-    @Test
-    void acceptRestrictions_Restriction_newUserAuthority_success() {
-        // given
-        Status pendingStatus = Status.builder().domain("RESTRICTION").code("PENDING").build();
-        Authority authority = Authority.builder().build();
-        ReflectionTestUtils.setField(authority, "authority_id", 1L);
-
-        RestrictionType restrictionType = RestrictionType.builder()
-                .duration(7)
-                .authority(authority)
-                .build();
-
-        RestrictionTarget restrictionTarget = RestrictionTarget.builder()
-                .restrictionType(restrictionType)
-                .status(pendingStatus)
-                .user(user1)
-                .build();
-        ReflectionTestUtils.setField(restrictionTarget, "restrictionTargetId", 1L);
-
-        Status completedStatus = Status.builder().domain("RESTRICTION").code("COMPLETED").build();
-        Status reportCompletedStatus = Status.builder().domain("REPORT").code("COMPLETED").build();
-        Status blurredStatus = Status.builder().domain("FEED").code("BLURRED").build();
-
-        when(restrictionTargetRepository.findById(1L)).thenReturn(Optional.of(restrictionTarget));
-        when(userAuthorityRepository.findByUserAndAuthority(user1, authority)).thenReturn(null);
-        when(statusManager.getStatus("RESTRICTION", "COMPLETED")).thenReturn(completedStatus);
-        when(statusManager.getStatus("REPORT", "COMPLETED")).thenReturn(reportCompletedStatus);
-        when(statusManager.getStatus("FEED", "BLURRED")).thenReturn(blurredStatus);
-        when(reportHistoryRepository.findByRestrictionTarget(restrictionTarget)).thenReturn(List.of(report1));
-
-        // when
-        reportService.acceptRestrictions(1L);
-
-        // then
-        verify(userAuthorityRepository).save(any(UserAuthority.class));
-        assertNotNull(restrictionTarget.getExpiresAt());
-        assertEquals(completedStatus, restrictionTarget.getStatus());
-        assertEquals(reportCompletedStatus, report1.getStatus());
-        assertEquals(blurredStatus, report1.getTransactionFeed().getStatus());
-        verify(transactionFeedRepository).saveAll(anyList());
-    }
-
-    @DisplayName("제재 승인_userAuthority에 제재 권한 내역 없는 경우 기존 권한 제재 연장")
-    @Test
-    void acceptRestrictions_Restriction_extendUserAuthority_success() {
-        // given
-        Status pendingStatus = Status.builder().domain("RESTRICTION").code("PENDING").build();
-        Authority authority = Authority.builder().build();
-        RestrictionType restrictionType = RestrictionType.builder()
-                .duration(7)
-                .authority(authority)
-                .build();
-        ReflectionTestUtils.setField(authority, "authority_id", 1L);
-
-        LocalDateTime previousExpiry = LocalDateTime.now();
-        UserAuthority userAuthority = UserAuthority.builder()
-                .user(user1)
-                .authority(authority)
-                .expiredAt(previousExpiry)
-                .build();
-
-        RestrictionTarget restrictionTarget = RestrictionTarget.builder()
-                .restrictionType(restrictionType)
-                .status(pendingStatus)
-                .user(user1)
-                .build();
-        ReflectionTestUtils.setField(restrictionTarget, "restrictionTargetId", 1L);
-
-        Status completedStatus = Status.builder().domain("RESTRICTION").code("COMPLETED").build();
-        Status reportCompletedStatus = Status.builder().domain("REPORT").code("COMPLETED").build();
-        Status blurredStatus = Status.builder().domain("FEED").code("BLURRED").build();
-
-        when(restrictionTargetRepository.findById(1L)).thenReturn(Optional.of(restrictionTarget));
-        when(userAuthorityRepository.findByUserAndAuthority(user1, authority)).thenReturn(userAuthority);
-        when(statusManager.getStatus("RESTRICTION", "COMPLETED")).thenReturn(completedStatus);
-        when(statusManager.getStatus("REPORT", "COMPLETED")).thenReturn(reportCompletedStatus);
-        when(statusManager.getStatus("FEED", "BLURRED")).thenReturn(blurredStatus);
-        when(reportHistoryRepository.findByRestrictionTarget(restrictionTarget)).thenReturn(List.of(report1));
-
-        // when
-        reportService.acceptRestrictions(1L);
-
-        // then
-        assertTrue(userAuthority.getExpiredAt().isAfter(previousExpiry));
-        assertEquals(completedStatus, restrictionTarget.getStatus());
-        assertEquals(reportCompletedStatus, report1.getStatus());
-        assertEquals(blurredStatus, report1.getTransactionFeed().getStatus());
-        verify(transactionFeedRepository).saveAll(anyList());
-    }
-
-    @DisplayName("제재 승인_실패(AlreadyProcessedRestriction)")
-    @Test
-    void acceptRestrictions_Fail_AlreadyProcessedRestriction() {
-        // given
-        Status acceptedStatus = Status.builder().domain("RESTRICTION").code("COMPLETED").build();
-        RestrictionType dummyRestrictionType = RestrictionType.builder()
-                .duration(7)
-                .build();
-        RestrictionTarget restrictionTarget = RestrictionTarget.builder()
-                .user(user1)
-                .status(acceptedStatus)
-                .restrictionType(dummyRestrictionType)
-                .build();
-        ReflectionTestUtils.setField(restrictionTarget, "restrictionTargetId", 1L);
-
-
-        when(restrictionTargetRepository.findById(1L)).thenReturn(Optional.of(restrictionTarget));
-        when(statusManager.getStatus("RESTRICTION", "COMPLETED")).thenReturn(acceptedStatus);
-
-        // when & then
-        assertThrows(AlreadyProcessedRestrictionException.class,
-                () -> reportService.acceptRestrictions(1L));
-    }
-
-    @DisplayName("제재 거절_성공")
-    @Test
-    void rejectRestrictions_success() {
-        // given
-        Status acceptedStatus = Status.builder().domain("RESTRICTION").code("COMPLETED").build();
-        Status rejectedStatus = Status.builder().domain("RESTRICTION").code("REJECTED").build();
-
-        RestrictionTarget restrictionTarget = RestrictionTarget.builder()
-                .user(user1)
-                .status(acceptedStatus)
-                .build();
-        ReflectionTestUtils.setField(restrictionTarget, "restrictionTargetId", 1L);
-
-        when(restrictionTargetRepository.findById(1L)).thenReturn(Optional.of(restrictionTarget));
-        when(statusManager.getStatus("RESTRICTION", "REJECTED"))
-                .thenReturn(rejectedStatus);
-        // when
-        reportService.rejectRestrictions(1L);
-
-        // then
-        assertEquals(rejectedStatus, restrictionTarget.getStatus());
-    }
-
-    @DisplayName("제재 거절_실패(AlreadyProcessedRestriction)")
-    @Test
-    void rejectRestrictions_Fail_AlreadyProcessedRestriction() {
-        // given
-        Status rejectedStatus = Status.builder().domain("RESTRICTION").code("REJECTED").build();
-
-        RestrictionTarget restrictionTarget = RestrictionTarget.builder()
-                .user(user1)
-                .status(rejectedStatus)
-                .build();
-        ReflectionTestUtils.setField(restrictionTarget, "restrictionTargetId", 1L);
-
-        when(restrictionTargetRepository.findById(1L)).thenReturn(Optional.of(restrictionTarget));
-        when(statusManager.getStatus("RESTRICTION", "REJECTED"))
-                .thenReturn(rejectedStatus);
-
-        // when & then
-        assertThrows(AlreadyProcessedRestrictionException.class,
-                () -> reportService.rejectRestrictions(1L));
+        verify(reportHistoryRepository).saveAll(anyList());
     }
 
 
@@ -628,53 +377,4 @@ class ReportServiceTest {
                 .isInstanceOf(ReportNotFoundException.class);
     }
 
-
-    @DisplayName("제재 id로 신고내역 조회_성공")
-    @Test
-    void getRestrictionReportHistory_Success() {
-        // given
-        Long restrictionId = 1L;
-
-        RestrictionTarget restrictionTarget = RestrictionTarget.builder().build();
-        ReflectionTestUtils.setField(restrictionTarget, "restrictionTargetId",restrictionId);
-
-        List<RestrictionReportResponseDto> mockResponse = List.of(
-                new RestrictionReportResponseDto(1L, "욕설 및 비속어 포함", "욕설입니다", LocalDateTime.now(), "AI 승인"),
-                new RestrictionReportResponseDto(2L, "욕설 및 비속어 포함", "또 욕설", LocalDateTime.now(), "관리자 승인")
-        );
-
-        when(restrictionTargetRepository.findById(restrictionId))
-                .thenReturn(Optional.of(restrictionTarget));
-
-        when(reportHistoryRepository.getRestrictionReportList(restrictionId))
-                .thenReturn(mockResponse);
-
-        // when
-        List<RestrictionReportResponseDto> result = reportService.getRestrictionReportHistory(restrictionId);
-
-        // then
-        assertThat(result.size()).isEqualTo(2);
-        assertThat(result.get(0).getReportId()).isEqualTo(1L);
-        assertThat(result.get(1).getStatus()).isEqualTo("관리자 승인");
-
-        verify(restrictionTargetRepository).findById(restrictionId);
-        verify(reportHistoryRepository).getRestrictionReportList(restrictionId);
-    }
-
-    @DisplayName("제재 id로 신고내역 조회_RestrictionNotFound")
-    @Test
-    void getRestrictionReportHistory_RestrictionNotFound() {
-        // given
-        Long restrictionId = 99L;
-
-        when(restrictionTargetRepository.findById(restrictionId))
-                .thenReturn(Optional.empty());
-
-        // when & then
-        assertThatThrownBy(() -> reportService.getRestrictionReportHistory(restrictionId))
-                .isInstanceOf(RestrictionTargetNotFoundException.class);
-
-        verify(restrictionTargetRepository).findById(restrictionId);
-        verify(reportHistoryRepository, never()).getRestrictionReportList(any());
-    }
 }
